Refactor Guitar filename to follow class module naming convention and update Guitar class import statements
