+++ prac_06/guitar.py
