+++ prac_06/guitar_test.py
@@ -1,4 +1,4 @@
-from prac_06.Guitar import Guitar
+from prac_06.guitar import Guitar
 
 # vintage guitar - 98 years
 gibson_l5_ces = Guitar("Gibson L-5 CES", 1922, 16035.40)
